added the combine function

+++ i2nca/tests/test_conv_pp.py
@@ -2,17 +2,18 @@ import unittest
 import os as os
 
 from m2aia import ImzMLReader
-from i2nca import convert_pp_to_pp_imzml, squeeze_pp_to_cp_imzml, squeeze_profile_to_pc_imzml, loc_max_preset
-
+from i2nca import convert_pp_to_pp_imzml, convert_pp_to_cp_imzml, convert_profile_to_pc_imzml, loc_max_preset
 
 
 def get_wdir(rel_path:str):
     return str(os.path.join(os.getcwd(), rel_path))
 
+def delete_output():
+    return False
 
-class TestConvTools(unittest.TestCase):
+class TestConvToolsPP(unittest.TestCase):
 
-    def test_conv_pp_to_pp_imzml(self):
+    def test_convert_pp_to_pp_imzml(self):
         # get data paths
         input_data = get_wdir(r"testdata\pp.imzML")
         output = get_wdir(r"tempfiles\pp")
@@ -27,26 +28,57 @@ class TestConvTools(unittest.TestCase):
 
         # check if m2aia parses new file
         I = ImzMLReader(result)
-        # cleanup temp files
-        os.remove(result)
 
-    def test_conv_pp_to_cp_imzml(self):
+        # cleanup temp files
+        if delete_output() == True:
+            os.remove(result)
+
+    def test_convert_pp_to_cp_imzml_def_param(self):
         # get data paths
         input_data = get_wdir(r"testdata\pp.imzML")
         output = get_wdir(r"tempfiles\cp")
 
         # convert data
-        squeeze_pp_to_cp_imzml(input_data, output)
+        convert_pp_to_cp_imzml(input_data, output)
 
         # expected result
         result = output + "_conv_output_cont_profile.imzML"
+        result2 = output + "_control_report_pp_to_cp.pdf"
 
         self.assertTrue(os.path.isfile(result))
+        self.assertTrue(os.path.isfile(result2))
 
         # check if m2aia parses new file
         I = ImzMLReader(result)
+
         # cleanup temp files
-        os.remove(result)
+        if delete_output() == True:
+            os.remove(result)
+            os.remove(result2)
+
+    def test_convert_pp_to_cp_imzml_coverage(self):
+        # get data paths
+        input_data = get_wdir(r"testdata\pp.imzML")
+        output = get_wdir(r"tempfiles\cp")
+        coverage = 0.5
+
+        # convert data
+        convert_pp_to_cp_imzml(input_data, output, coverage)
+
+        # expected result
+        result = output + "_conv_output_cont_profile.imzML"
+        result2 = output + "_control_report_pp_to_cp.pdf"
+
+        self.assertTrue(os.path.isfile(result))
+        self.assertTrue(os.path.isfile(result2))
+
+        # check if m2aia parses new file
+        I = ImzMLReader(result)
+
+        # cleanup temp files
+        if delete_output() == True:
+            os.remove(result)
+            os.remove(result2)
 
     def test_conv_pp_to_pc_imzml(self):
         # get data paths
@@ -54,35 +86,22 @@ class TestConvTools(unittest.TestCase):
         output = get_wdir(r"tempfiles\pc")
 
         # convert data
-        squeeze_profile_to_pc_imzml(input_data, output, loc_max_preset)
+        convert_profile_to_pc_imzml(input_data, output, loc_max_preset)
 
         # expected result
         result = output + "_conv_output_proc_centroid.imzML"
+        result2 = output + "_control_report_prof_to_pc.pdf"
 
         self.assertTrue(os.path.isfile(result))
+        self.assertTrue(os.path.isfile(result2))
 
         # check if m2aia parses new file
         I = ImzMLReader(result)
+
         # cleanup temp files
-        os.remove(result)
-
-    def test_conv_cp_to_pc_imzml(self):
-        # get data paths
-        input_data = get_wdir(r"testdata\cp.imzML")
-        output = get_wdir(r"tempfiles\pc")
-
-        # convert data
-        squeeze_profile_to_pc_imzml(input_data, output, loc_max_preset)
-
-        # expected result
-        result = output + "_conv_output_proc_centroid.imzML"
-
-        self.assertTrue(os.path.isfile(result))
-
-        # check if m2aia parses new file
-        I = ImzMLReader(result)
-        # cleanup temp files
-        os.remove(result)
+        if delete_output() == True:
+            os.remove(result)
+            os.remove(result2)
 
 
 if __name__ == "__main__":
